LayoutTests/http/wpt/navigation-api/transition-promises-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: UnknownError: Uncaught Error:

Harness Error (FAIL), message = Unhandled rejection: Uncaught Error:

PASS NavigationTransition finished promise is fulfilled on success
PASS NavigationTransition finished promise is rejected on error
